+++ service/database/change-name.go
@@ -2,7 +2,7 @@ package database
 
 import "strings"
 
-func (db *appdbimpl) ChangeName(username string, id int) error {
+func (db *appdbimpl) ChangeName(username string, id uint64) error {
 	_, err := db.c.Exec("UPDATE Users SET name = ? WHERE id = ?", username, id)
 
 	if err != nil {
